Refactor query plans controller and presenter
Controller and presenter have been separated and moved into their corresponding controller/presenter folders.
